move resizer and writer code to their own file
