+++ Makefile
@@ -18,7 +18,7 @@ venv-lint-test: ## [Continuous integration]
 
 test: ## [Local development] Run unit tests
 	rm -rf tests/test_folder/
-	python -m pytest -v --cov=img2dataset --cov-report term-missing --cov-fail-under .45 tests
+	python -m pytest -x -v --cov=img2dataset --cov-report term-missing --cov-fail-under .45 tests
 
 .PHONY: help
